Add deep-linking search feature

+++ plugins/pm_filter.py
@@ -2585,10 +2585,14 @@ async def auto_filter(client, name, msg, reply_msg, ai_search, spoll=False):
         else:
             return
     else:
-        message = msg.message.reply_to_message  # msg will be callback query
+        if isinstance(msg, CallbackQuery):
+            message = msg.message.reply_to_message  # msg will be callback query
+        else:
+            message = msg
         search, files, offset, total_results = spoll
         settings = await get_settings(message.chat.id)
-        await msg.message.delete()
+        if isinstance(msg, CallbackQuery):
+            await msg.message.delete()
     pre = 'filep' if settings['file_secure'] else 'file'
     key = f"{message.chat.id}-{message.id}"
     req = message.from_user.id if message.from_user else 0
